Only style parent if it hasn't been styled yet, otherwise style child
This adds some checking and methods to do so. Essentially, each $$Barbershop instance has a WeakReference<T> field that holds a reference to the last target styled (which should fall through from parent to children). If a style() method is called, it will only call the super's style method if the parent hasn't been styled yet, otherwise it proceeds as normal and updates the current last target styled with itself. This however isn't threadsafe